refactored Appbar
