+++ client/src/components/Home/index.js
@@ -1,14 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { MuiThemeProvider, createTheme, styled } from "@material-ui/core/styles";
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
-import Box from "@material-ui/core/Box";
+import { Grid, Typography, Box, Button } from "@material-ui/core";
 import history from '../Navigation/history';
-import AppBar from '@material-ui/core/AppBar';
-import Container from '@material-ui/core/Container';
-import Toolbar from '@material-ui/core/Toolbar';
-import Button from '@material-ui/core/Button';
+import { AppMenuBar } from '../AppMenuBar';
 
 //Dev mode
 const serverURL = ""; //enable for dev mode
@@ -103,75 +98,7 @@ class Home extends Component {
             backgroundSize: "cover"
           }}
         >
-          <AppBar position="static">
-            <Container maxWidth="xl">
-              <Toolbar disableGutters>
-                <Button
-                  key='1'
-                  onClick={() => history.push('/')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Home
-                </Button>
-                <Button
-                  key='2'
-                  onClick={() => history.push('/SearchSchedule')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Schedule
-                </Button>
-                <Button
-                  key='3'
-                  onClick={() => history.push('/SignIn')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Sign In
-                </Button>
-                <Button
-                  key='9'
-                  onClick={() => history.push('/SignUp')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Sign Up
-                </Button>
-                <Button
-                  key='4'
-                  onClick={() => history.push('/MyProfile')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  MyProfile
-                </Button>
-                <Button
-                  key='7'
-                  onClick={() => history.push('/FAQ')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  FAQ
-                </Button>
-                <Button
-                  key='8'
-                  onClick={() => history.push('/Location')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Location
-                </Button>
-                <Button
-                  key='10'
-                  onClick={() => history.push('/OtherReviews')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Reviews
-                </Button>
-                <Button
-                  key='11'
-                  onClick={() => history.push('/Annoucements')}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  Annoucements
-                </Button>
-              </Toolbar>
-            </Container>
-          </AppBar>
+          <AppMenuBar />
           <MainGridContainer
             container
             spacing={5}
